risks/alert_app/requirements.txt
google-cloud~=0.34
flask~=3.0.0
google-cloud-pubsub~=2.19
google-cloud-storage~=2.10.0
psycopg2~=2.9.9